Add test to check for changes to the compiler passes
[reviewed by tom]

This is a simple check to test if the compiler passes have changed for the
purpose of notifying us when we have to update the compiler performance
testing. 

Unfortunately it's hard to automate adding/reordering/removing passes. At least
for anything even a little complex. How do you automatically tell the
difference between if a pass was removed and a new one was added vs simple
renaming.

When a pass is changed the compilerPerformance.perfkeys has to be updated. The
.dat files have to be updated and the .graph files have to be updated.

For now I think it makes sense just give me a heads up when a pass changes and
I'll take care of it. Passes don't change often so this shouldn't be a big
deal.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22898 3a8e244f-b0f2-452b-bcba-4c88e055c3ca