Reduce the number of (unnecessary) formal temps inserted by the compiler
[reviewed by hilde]

At some point about six months ago, I found myself trying to figure
out why we insert so many formal temps (temps for formal arguments to
functions) in the generated code; in looking into it, I was
overwhelmed by the logic in function resolution that determines
whether or not we do.  My conjecture was that such temps should be
required in only a very small number of cases -- those requiring
copies on the way into the procedure that were not handled by the
back-end function calls themselves.

To that end, I tried ripping out all the logic and replacing it with
something far simpler: Insert a formal temp if the argument is 'inout'
or 'out' (because we pass by reference in these cases, so we need an
explicit copy); or for 'in or 'const in' intents if the type is one
for which we can't rely on C to do the copy for us.

This new logic worked surprisingly well, except that it unearthed some
pre-existing faults in the implementation that inserting formal temps
had previously been papering over.  The main one of those was fixed by
my recent commit relating to remote value forwarding.  This commit
addresses a few other smaller ones.

The benefits of this commit are: Simpler function resolution logic
w.r.t. formal temps; fewer formal temps inserted, reducing the size of
the AST (some of these were already eliminated before codegen by later
passes like copy propagation; others made it through to codegen and
now don't); fewer copies for routines that take record and tuple
arguments where formal temps had been inserted previously.

Overall, this commit also reduces the number of memory leaks and total
amount of memory allocated/freed in the examples directory.  However,
it does add one new (to me) leak case -- for some cases of zippered
forall iteration, the lack of a formal temp for the 'followThis'
argument results in an iterator class being leaked.  In talking with
Tom, he indicated that he had not done much to plug leaks for
iterators and was not surprised that there would be fragility here.
The overall win seemed worth the loss.  I do have a hack (currently
commented out) in this commit in functionResolution.cpp which
special-cases followThis arguments and can be used to plug cases of
this leak that I've seen, if we choose to do this.

With respect to "unearthing pre-existing faults", the main one worth
calling out is one related to function resolution of dynamically
dispatched methods.  The way in which I saw this show up is as
follows:

An array's copy code makes calls to here.id/locale.chpl_id() while it
is reasoning about bulk copy optimizations.  If:

(a) a code contains no calls to here.id and 

(b) these copy routines are not introduced into the user's code prior
to function resolution and 

(c) calls to these copies are introduced
after function resolution

then the compiler incorrectly decides that the chpl_id() call on the
locale model should be statically dispatched to the abstract base
class type's "pure virtual" function which throws an error rather than
properly dynamically dispatching the call.  Needless to say, this is
problematic.  The workaround I used to get around this issue was to
introduce a call to here.id in user code -- specifically in a new
internal module ChapelDynDispHack.chpl.  If this call is removed,
tests like the following will fail:

        test/classes/bradc/arrayInClass/arrayDomInClassRecord-simpler.chpl

I'm 95% certain that this is a pre-existing problem with function
resolution, yet it's one that I was unable to reproduce in a simpler
program.  Again, the reason it's showing up now is that I've reduced
the number of temps inserted, resulting in fewer copies in user code,
resulting in fewer calls to here.id/chpl_id() in user code (in cases
like the above, none), and so the copy function only gets resolved
in the special resolveAutoCopies() pass of functionResolution.cpp;
at which point it is incorrectly resolved statically rather than
dynamically.


The main other problematic case that I'm aware of is

  test/multilocale/diten/needMultiLocales/remoteStringTuple.chpl

which now fails on multiple locales due to passing a non-local string
to an extern routine.  My hypothesis was that it would be better to
wrestle with this case after strings are converted into records and
Sung agreed, so we'll let this one fail for now.


Here are some additional file-by-file notes:

functionResolution.cpp: Contains the changes to formal_temp insertion
logic as described above.

symbol.cpp: Fixes one of the formerly hidden sins, by flagging that
sync variables passed by blank intent are not constant.

ChapelBase.chpl: chpldev_refToString() used to be a hack that relied
on 'inout' intents; now that we have ref intents, it can (and, without
formal temps, needs to be) written more properly using a 'ref' intent.

ChapelStandard.chpl and ChapelDynDispHack.chpl: These are the module
changes that contain the hack relating to dynamic dispatch of here.id
noted above.

printModuleInitOrder.good: changes to reflect the new
ChapelDynDispatchHack.chpl module described just above.

refcnt.good: now that we don't put in formal temps for records, the
number of constructor/destructor calls in this program are now
significantly reduced (presumably a good thing), so the .good is
updated to reflect that.

test/*/*.chpl: removing formal temps revealed that some temps were
changing records (passed by 'const' intent by default) illegally.  Our
const checking isn't good enough to flag these cases; and the formal
temps were preventing the modifications from showing up back at the
callsite; once they were removed, the problems became evident.

arrayDomInClassRecord-illegal.chpl, recordpass6-illegal.chpl: These
futures capture the case that assignment to records passed by blank
intent should flag an illegal assignment error, but don't currently.
Previously, because of the formal temps that were inserted, such local
assignments were permitted and didn't reflect back to the callsite;
now that we're squashing formal temps, the illegal modifications were
reflected back at the callsite.

arrayDomInClassRecord-simpler.chpl: As noted above, this is the
simplified version of the code that causes the pure virtual function
error when there aren't user-level calls to here.id.






git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22900 3a8e244f-b0f2-452b-bcba-4c88e055c3ca