Enable launch-gasnetrun_ibv to pass CHPL_ environment variables
[reviewed by Brad and Rafael Larrosa Jiménez]

Several important environment variables, including
CHPL_RT_NUM_THREADS_PER_LOCALE, need to be passed on to the spawned jobs.
Unfortunately, with the gasnetrun_ibv launcher using SSH (vs mpirun), these
important environment variables are not passed on.  The gasnetrun_ibv.pl
script knows how to accept a -E option which is followed by a comma-separated
list of environment variables that should be forwarded. This patch enables
the use of this -E option and generates a list of environment variables that
includes all environment variables. It does so with a hopefully more
general-purpose function, chpl_get_enviro_keys, which returns a string
containing environment variable names that should be forwarded, separated by
the passed character. 




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22912 3a8e244f-b0f2-452b-bcba-4c88e055c3ca