Updating cyclic comm counts to reflect formal temp-based improvements
The comm counts for tests got significantly better with the formal
temp improvements yesterday, so I'm updating their .goods to reflect
the new scores.  reduce and assign in this directory also got better
compared to where they were, but still aren't back to where they were
prior to the existing regression for them, so I'm leaving them alone
for now.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22914 3a8e244f-b0f2-452b-bcba-4c88e055c3ca