Disable PSHM in GASNet builds
[reviewed by Greg]

Since Chapel uses threads to access multiple cores/sockets
on a machine, we don't need GASNet's shared-memory
communication scheme for different processes on the same node.
(In fact it does not work on some configurations, including
 some Infiniband and some Cray systems. PSHM was enabled by
 default in GASNet at some point).

PSHM should be re-enabled on a case-by-case basis for
targets on which it is useful. However, such targets
seem to be the exception at this point, since Chapel
will make use of on-node parallelism with threads vs.
processes.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22915 3a8e244f-b0f2-452b-bcba-4c88e055c3ca