Minor update to the intrinsic atomics
[reviewed by sung]

This removes the fake return that I added in r22904 to prevent gcc from
complaining about a computed value not being used. To do this we have to cast
the bools to ints for the intrinsic functions. I'm not really sure why gcc
complained for the bool version and not the int but it's the right change since
the intrinsic operations are technically only supported for integral variables
anyways. 

I also added a few notes at the bottom for things that will matter when we add
an implementation that uses C11 atomics.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22923 3a8e244f-b0f2-452b-bcba-4c88e055c3ca