Fix a --no-local regression related to export functions
[reviewed by hilde]

In r22900, I introduced a regression that manifested itself in
test/users/ferguson/ddata/cToChapelArray.chpl in which an exported
function started having wide arguments.  The cause of this was that,
prior to my commit, the presences of a formal temp caused the compiler
to narrow the wide argument type, whereas the lack of such a formal
temp caused it to think it could not do so.

The approach that I wanted to take here was to have the parallel.cpp
pass simply not widen argument types by extending the following
conditional to look for FLAG_EXPORT (or, more precisely,
FLAG_LOCAL_ARGS which is set when FLAG_EXPORT is true):


    // Widen all variables,
    // and all arguments of functions not marked "extern".
    if (isVarSymbol(def->sym) || isArgSymbol(def->sym))
    {
      if (Type* wide = wideClassMap.get(def->sym->type))
        if (isVarSymbol(def->sym) ||
            !def->parentSymbol->hasFlag(FLAG_EXTERN))
          def->sym->type = wide;
    }
  }

Unfortunately, this approach didn't work.  In particular, export
functions that took narrow arguments but passed them to routines
expecting wide arguments (like chpldev_taskTable_add()) never widened
the arguments in question.  The reason seems to be (thanks to Tom's
analysis) that we don't widen references simply because a narrow is
being passed to a wide; we only widen them when a wide thing has
already been narrowed.  This doesn't make sense to me and seems like
something that should be addressed in the future.  I.e., imagine a
pass that says "This thing is narrow, but being passed to a routine
that wants a wide; let's widen it!"

So, instead, I went into the narrowSym() routine and found that, for
the case in question, the argument fell all the way to the bottom of
the routine where it decides to make the argument wide, having not
fallen into any of the other cases along the way.  This seems
suspicious to me (as noted in the comments I added), and in any case
was able to be fixed by simply putting in a new conditional that
checked to see if FLAG_LOCAL_ARGS was set and avoided forcing the
thing to be widened if so.  This fixed the test in question.

It may also have the side benefit of removing some other unnecessary
widenings in the body of a routine, but since it only applies to
EXPORT/EXTERN functions (and those with an explicit "local args"
pragma), I don't expect to see returns on any key performance tests.
If we do, it'll simply be a happy surprise.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22925 3a8e244f-b0f2-452b-bcba-4c88e055c3ca