Add capability to annotate performance graphs
This patch adds the ability to place annotations on our performance graphs.
To do this, add an entry into $CHPL_HOME/test/ANNOTATIONS.yaml in the
format below (pulled from comments in ANNOTATIONS.yaml):

group:
  mm/dd/yy:
    - example annotation 1
    - example annotation 2
  mm/dd/yy:
    - text: example annotation 3
      host: [hostname1, hostname2]

all:
  mm/dd/yy:
    - example annotation 5

where group is the name of the .graph file that attached set of annotations
should apply to. When the long form of an annotation is used (when text and
host are specified) the annotation is only added to the graphs if the current
hostname is in the host list. 'all' is a special group where any annotations
are added to to every single graph. One limitation of dygraphs is that if you
try and place an annotation on a date that has no data (tests didn't run or
all failed) it will not be displayed.

To enable this functionality, the PyYaml[0] package is required. It can be
installed as described on their homepage or via `pip install pyyaml` (which I
would recommend). If the module is not found, annotations will simply not be
displayed.

[0] http://pyyaml.org/wiki/PyYAML


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22926 3a8e244f-b0f2-452b-bcba-4c88e055c3ca