Stopgap solution to the copyPropagation hang discovered by Kyle. [reviewed by eronagha]
I traced the reason for the hang in copy propagation in Kyle's test code to an
assumption in the textbook algorithm that is being violated by our basic block
analysis.  The algorithm assumes that only the entry block has no
predecessors.

In global copy propagation, the initial state of the IN set is set to all false
for the entry block and all true for all other blocks.  But in a general
solution, the initial state of the IN set should be zeroes for each block that
has no predecessors (and all ones for the rest).

An alternate (and better) solution would be to ensure that basic block analysis
attaches at least one predecessor to every interior block.  In the mean time,
this modification to the setup for GCP ensures that it will run without
hanging.

[tested on / (std)]
[tested on copy-hang (with Kyle's module changes)]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22928 3a8e244f-b0f2-452b-bcba-4c88e055c3ca