Re-enable GASNet pshm in circumstances where it's helpful.
[Reviewed by Sung.]

r22915 disabled pshm in all cases, but sometimes it's actually helpful.

With pshm enabled, when using the udp or smp conduits with seg=fast or
seg=large GASNet will communicate between processes on the same system
node via shared memory.  We have multiple Chapel processes on a single
system node whenever we run multilocale programs using local spawning,
including for nightly testing.  In the particular case of gasnet-fast
local testing, using pshm reduces the testing time by 20%, and more
importantly, the reduction in overhead helps to expose race conditions.
So, enable pshm for that case and leave it disabled elsewhere.  (We
actually also enable it for seg=large, on the offhand chance we'll test
that someday.  The pshm support is not usable with seg=everything.)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22929 3a8e244f-b0f2-452b-bcba-4c88e055c3ca