Use assign-to-ref for array assignment. [reviwed by lydia, vass, bradc]
Change array assignment to use the new signature, where the LHS has ref intent
and the return type is void:
  proc =(ref a:[], b) : void;

This patch includes three elements: A factoring of the implementation of
proc =() for arrays, replacement of the old signature with the new, and fixing a
resulting bug associated with yielding arrays from iterators.

This patch also incidentally changes the way some warnings are issued from 7
tests.  This is due to the refactoring of array assignment interacting with a
weak implementation of reissueCompilerWarning() in functionResolution.cpp.
This version of the patch accepts the change in behavior and defers the
reworking of reissueCompilerWarning().

Factoring array assignment made it easier to partition the work of converting
to the new signature and to ensure its correctness.  Moving to the new
signature was then fairly trivial, except for the bug it turned up in iterators
returning arrays by value.

The bug appeared because the fields in an iterator record used to represent
local variables are never initialized, and the code for initializing arrays in
iterator bodies is explicitly removed.  In an iterator, the yield statement is
replaced by an assignment at some point.  Since the new version of assignment
expects the LHS to be a fully-initialized array object, the program fails at
runtime with a segfault -- due to the attempt to access memory through an
uninitialized pointer.

The solution for the bug was return arrays using var intent even in the case
where the return value is specified and var intent is not specified explicitly.
The code to perform this was added at line 491 in normalize.cpp.

After that was done, I noticed that the code for explicitly removing array
initialization from iterators became dead code, so I deleted it.  (See the "if
(removeExpr)" conditional on line 6755 (old) of functionResolution.cpp.)  


DETAILS

Index: modules/internal/ChapelArray.chpl
Factored the long proc=() with an array as its LHS, to eliminate the code
repetition related to the bulk copy implementation, and also to separate the
cases where the RHS is of different types (pushing these up into "where"
clauses).
Adjusted the indentation of the body of tuple-to-array assignment.
Replaced the call to array assignment in the array initCopy function with
direct transfers instead.

Index: compiler/include/iterator.h
Index: compiler/AST/iterator.cpp
The code for removing initialization of arrays in iterators was deleted, so
the list of iteratorsWithRemovedRetInits will always be empty.
The code for removing such initializations was also deleted.

Index: compiler/passes/normalize.cpp
Added a predicate to test if a given return type expression indicates that it
is an array type.  This predicate is used to test if an iterator is yielding an
array.  The test is used before resolution, so we can't simply look to see if
it's an array type.
If an iterator is yielding an array by value, we change the iterator to yield
it by ref (var) intent instead.  This doesn't change the language semantics,
but does avoid the bug involving accessing uninitialized memory at runtime.

Index: compiler/resolution/functionResolution.cpp
Deleted the code related to explicitly removing array initialization code from
iterators.

[tested on / (std)]
[tested on / (--no-local)]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22933 3a8e244f-b0f2-452b-bcba-4c88e055c3ca