Enabling formal temp 'followThis' hack
As mentioned in the commit for r22900, with the removal of formal
temps, leaks were increased for cases of zippered forall loops once
formal temps were no longer inserted for the 'followThis' arguments.
I'm enabling the hack that I'd inserted but left commented out in the
commit to see how it impacts overall testing (I've done correctness
testing over the whole suite with it and examples memory leaks testing
which it helps with, but not broader memory leaks or performance
testing).  If things look bad in some way tomorrow, I'll back it back
out.  If things look good, we should discuss as a group whether we're
OK with keeping it in.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22935 3a8e244f-b0f2-452b-bcba-4c88e055c3ca