Add release over release performance graphs for chap04
[reviewed by thomas] 

Typically we have nightly performance graphs that show the state of the
performance tests with the compiler that existed on that night. Release over
release performance shows how the current performance tests compare amongst
various releases of the compiler as well as the most recent nightly. This
eliminates changes to the performance tests and shows us just the comparison of
the compiler from release to release and with the current trunk. 

We basically have the historical data tucked away (which has to be manually
updated when the performance tests are changed) and every night we splice the
historical data with the nights performance results. We then create new graphs
and move them over to sourceforge. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22939 3a8e244f-b0f2-452b-bcba-4c88e055c3ca