Update release over release performance graphs
In the previous commit release over release performance graphs were added.
That implementation found all the .dat files in the folder containing the
historical data and found the corresponding files for nightly. The problem with
this is that not all tests compile for previous releases. What that meant is
that some .dat files were missing which screwed up genGraphs. So rather than
just some series missing from graphs, the wrong series name was used. For
instance meteor has 4 series but it was showing the data for
meteor-parallel-alt but had a series name of meteor. 

To fix this problem this patch instead of looking for the .dat files in the
historical folder it finds all the .dat file in the nightly run that have been
modified in the last 48 hours. Then if there is no corresponding historical
file we just take the data from the nightly file. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22942 3a8e244f-b0f2-452b-bcba-4c88e055c3ca