Reformat/improve wording of internal failures/errors
Prior to this patch, internal errors/failures printed
the following for a user:

        'error: internal failure CAL123'

and the following for a developer:

       'error: <message intended for developers>'

Vass, in internal-error-message.future, suggested (reasonably) that
the developer's error message should say something indicating that it
is an internal failure/error as well -- otherwise, it is difficult to
distinguish from a typical error for an end-user.

To that end, after some negotiation, we settled on making the error
messages be the following for an end-user:

       'internal error: CAL123'

and the following for a developer:

       'internal error: <message intended for developers>'


The change itself was simple (a 5-minute fix (TM)); the work was
updating .prediffs, .bads, and .goods to reflect the new format.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22960 3a8e244f-b0f2-452b-bcba-4c88e055c3ca