Update nbody_fullreduction.future
This test doesn't fail in the way described anymore.  It gets a nil
dereference.  The original text said that this would happen if some
for loop was removed, but I don't know which one.  In any case, it
seems that the original problem might be gone, so I updated the future
bug description to be a little more specific, added a .bad file and
prediff.

Also, updated to the new (ha!) zip() syntax (the test has been failing
compilation since zip() was introduced).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22961 3a8e244f-b0f2-452b-bcba-4c88e055c3ca