Added simple const-ness checking for records.
[reviewed by Tom]

With this commit, the following exprs are treated as consts,
except as described later:

* accesses of fields of const records/unions, and
* accesses of const fields of records/unions/classes.

By "access" I mean invoking the field accessor function,
whether user-defined or compiler-generated.

Given that, the usual const-checking machinery applies,
in particular assignments and passing to lvalue-requiring
formals (ref, out, ...) will be flagged as errors.

Also, an assignment to a record with const field(s)
is disallowed. This is because such an assignment would
include potentially modifying the record's const field(s).
However, the following are allowed:
* initializing such a record (see exception below), and
* assigning to such a record using a custom assignment operator
  that does not modify any const fields.

Exception: modifying const fields of a record is allowed:
- in a constructor for the record,
- in an initialize() method invoked on the record as the receiver,
- in a method invoked directly from a constructor or initialize(),
  again with such record as the receiver.

The motivation for this exception is to enable records with constant
fields to be constructed, given that currently the compiler currently
does not distinguish between assignment and initialization.
At the moment we do not have UMM worked out and implemented,
so the above is somewhat of a heuristic.


IMPLEMENTATION

A typical field access "MYRECORD.CONSTFIELD = 333333" looks like this:

      (1603 DefExpr 'call_tmp[1601]:_ref(int(64))[3159]')
      (1604 CallExpr move
        (1605 SymExpr 'call_tmp[1601]:_ref(int(64))[3159]')
        (861 CallExpr
          (3817 SymExpr 'fn CONSTFIELD[2221]:_ref(int(64))[3159]')
          (1600 SymExpr '_mt[188]:_MT[186]')
          (1369 SymExpr 'MYRECORD[854]:RECTYPE[751]')))
      (1607 CallExpr move
        (1606 SymExpr 'call_tmp[1601]:_ref(int(64))[3159]')
        (867 CallExpr
          (3818 SymExpr 'fn =[3571]:int(64)[70]')
          (1602 SymExpr 'call_tmp[1601]:_ref(int(64))[3159]')
          (866 SymExpr 333333 '_literal_41[864]:int(64)[70]'))))

where 'fn CONSTFIELD[2221]' is a field accessor function.

To detect that the assignment 867 is illegal, I do the following:

* Mark fn CONSTFIELD with FLAG_FIELD_ACCESSOR and FLAG_REF_TO_CONST
(if it accesses a const field).

* In resolveNormalCall(), check if the call is to a FLAG_FIELD_ACCESSOR
function. If so, mark the resulting call_tmp (1601 in this case) with
FLAG_REF_TO_CONST if:
 - the callee is also FLAG_REF_TO_CONST, or
 - the record being accessed (854 in this case) is FLAG_CONST.

Forego this bullet if we are in a constructor or initialize()
or a method invoked directly from either, and are accessing a
FLAG_ARG_THIS symbol. This heuristically allows
initialization of const fields during construction - given
that currently AST does not distinguish between initialization
and assignment.

* Expand the constness checking, esp. isLegalLvalueActualArg(),
to also check FLAG_REF_TO_CONST.

* When assignment to a record with const field(s) is detected,
the error comes from assigning to such a field within the record
assignment operator. For a compiler-generated assignment, this
error has the record type's line number, which is not helpful.
So instead when reporting such an error I shift the blame to
the caller of the record assignment.

Unfortunately such an error is still not always helpful. In
particular, e.g. in studies/madness/.../FTree.chpl, we got
a record type with const fields and an associative domain
over such a arecord. When the domain is resized (e.g. dom += elem),
domain indices get reassigned, hitting an error on assigning
records with const fields. The error message does not point
to that.

Also, in optimizations/bulkcomm/asenjo/ptransDR/v*/ptrans.chpl,
something invokes domain assignment, which invokes
dsiReallocate() on the domain's _arrs. So BaseArr.dsiReallocate()
is resolved, then addToVirtualMaps() resolves dsiReallocate
for all array types, including DefaultRectangular, which
re-assigns arrays over localInfo - a record declared in ptrans.chpl
with const fields. This triggers the same error as in the previous
paragraph, however the error message is unhelpful even with
--print-callstack-on-error.


CHAPEL SOURCE CODE CHANGES

Here are some highlights of what I had to change:

* In LocaleModel.chpl and a couple of other places, we have:

  class LocaleModel : AbstractLocaleModel {
    const callStackSize: size_t;
    ...
    proc LocaleModel() { // constructor
      ...
      init();
    }
    ...
    proc init() { ... callStackSize = chpl_task_getCallStackSize();  ... }
  }

The modification of callStackSize in init() is technically invalid,
so I changed the constructor to pass callStackSize to init() by ref.
Similarly for a couple of other LocaleModel's const fields.

Ideally we would not require users to change that.

* As mentioned above, (a) modifications of an associative domain or
(b) assignment to a rectangular domain or (c) assignment to array
elements prevent the record type from having const fields. For (a),
that's the type the domain is over, for (b,c) it's the element type of
an array (for (b) - over such a domain).

I changed 'const' to 'var' fields in these cases. However,
we need to figure out how to keep fields const.

* BlockDom.whole went from 'const' to 'var' - it gets assigned
upon dsiReprivatize. Ditto for Cyclic and BlockCyclic.

* studies/madness boldly and incorrectly declared Function.f
as const, then assigned to it here and there. I switched f to var.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22963 3a8e244f-b0f2-452b-bcba-4c88e055c3ca