Fix cygwin sensitivity to spaces in usernames
[reviewed by mppf, vass]

As mentioned on chapel-bugs this week, a user was having a 32-bit
cygwin compilation failure which turned out to be due to having a
space in his username.  Turns out that the commands that we use to
make and remove the temp directory that we create by default (and
maybe open files in that directory) are open to unprotected spaces in
the filename, due to our use of the system() command (I believe).

I initially looked into a change to protect those spaces with
backslashes before settling on this approach: compress the spaces out
of the username before even creating the /tmp directory name.  It's
slightly simpler and shouldn't ruin the goal of creating a reasonably
unique /tmp directory name.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22969 3a8e244f-b0f2-452b-bcba-4c88e055c3ca