Fix a weird bug with nohup.
[trivial and not reviewed]

When using nohup to run a Chapel-compiled program
on my system, I was getting an error from file.reader()
with path /dev/null because the standard input stream 0
was a file that was not readable. That makes no sense
at all, but in order to allow programs to start up in
this strange environment, this patch makes QIO always 
consider file descriptor for stdin to be readable and 
the file descriptors for stdout/stderr to be writeable. 
 
If there is a e.g. a stdin that is not readable, there 
will be a fatal error later on in the program if it tries 
to use stdin. Of course, in the common case with nohup, 
the program would not use stdin anyway. 

tested with test/io.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22971 3a8e244f-b0f2-452b-bcba-4c88e055c3ca