Update comm counts following r22674.
Following the conversion of most assignments to use 'ref' intent on the LHS
operand, the "get" count in some comm diagnostics increased.  This is expected,
since some fundamental types or records that were being passed by value as the
left operand of assignment are now getting passed by ref.  The get counts in
the three differing .bad and .good files have been updated to reflect this new
behavior.

Passing the LHS by ref entails an extra get in assignments where the LHS is on
a remote node.  That is to obtain the address of the LHS operand.  Previously,
(due to the presence of a "wrapper" move, the RHS would be moved into local
temp, and then a put operation would be performed to copy the value of the
local temp into the remote destination.  All of those actions still happen, so
the only visible difference is the extra get to obtain the destination address.

[tested on these three test cases (gasnet-everything)]
[reviewed by diten]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22972 3a8e244f-b0f2-452b-bcba-4c88e055c3ca