Clean up issues with performance test verification
[Reviewed by Lydia]

Fasta-lines currently fails due to an OOM error and prints out a very
large amount of junk, drastically increasing the size of the log files.
To fix this, the test output is either

1) printed in its entirety when it is less than 2000 characters

or

2) the first and last 1000 characters of the output are printed

In addition, the test output is run though a filter to replace any
non-printable characters with a ~. This is useful as some of our tests
output binary data which can cause issues when viewed with a text
editor. Also, any performance test that fails or times out is will no


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22979 3a8e244f-b0f2-452b-bcba-4c88e055c3ca