Applying a fix to tuple-to-extern
[reviewed by mppf]

Sung pointed me to this future the other day and asked what I thought
about it.  It seemed to me like a case of trying to work too hard to
reverse engineer stuff under Chapel's hood, similar to what we used
to suggest people do to get to the guts of an array; happily, changing
to the much more straightforward approach worked fine.  I.e., pretend
the external function is taking a tuple, and pass the tuple.  Updated
the test and am retiring the future.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22987 3a8e244f-b0f2-452b-bcba-4c88e055c3ca