Add --static flag for performance testing
[reviewed by Kyle]

This commit adds --static to the compopts for all performance tests.

For some reason, r22900 caused the fasta performance test to slow down
by ~15%.  Upon investigation, it seems that the slow down was only
seen when using dynamic linking and depended on the file system on
which Chapel installation was located (local physical disk).  It's
still a bit of a mystery to me as to how this could happen, but in
general it seems like you could get more stable timings using static
linking, especially for programs using unix time to get numbers.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22992 3a8e244f-b0f2-452b-bcba-4c88e055c3ca