Add slurm launcher
[reviewed by thomas]

This adds a slurm launcher with the intent of supporting native slurm on some
upcoming XC30's. That said it will work with non-xc30 machines running slurm. 

This launcher is very similar to the slurm_gasnetrun_ibv launcher. It has some
options removed and others added and had some cleanup done but it's similar.
The primary difference is that this launcher uses srun to launch a job instead
of gasnetrun. The launcher works with both interactive and batch jobs. 

For interactive jobs: 
  export CHPL_LAUNCHER_WALLTIME=00:15:00 (or however long you want) 
  for use with start_test: 
    CHPL_LAUNCHER_TIMEOUT=slurm
    pass --nostdinredirect to start test 


For batch jobs: 

  export CHPL_LAUNCHER_USE_SBATCH=true
  export CHPL_LAUNCHER_WALLTIME=00:15:00 (or however long you want, for batch
  jobs the walltime is not required and defaults to an hour on most systems)


To use the launcher with start_test either for batch or interactive there is a
prediff that is required. For interactive jobs all the prediff does is remove
the echos from salloc. For batch jobs it's a bit trickier. Start test really
just fires off the batch job submission. The prediff waits for the job to
complete by polling the status of squeue and figuring out if a job is completed
or failed and handles it accordingly. 

This launcher probably doesn't have the optimal options set (not that that's
different than most of the launchers) but it does complete testing on
release/examples with no problems. 

This patch also adds svn:ignore to the launcher gen directories that were
missing them. 

Some helpful links (especially for figuring out what the options do): 
  https://computing.llnl.gov/linux/slurm/sbatch.html
  https://computing.llnl.gov/linux/slurm/salloc.html
  https://computing.llnl.gov/linux/slurm/srun.html

Todos(the only near term one is probably any chplenv changes): 
 - Adding a help printout
 - Figuring out when to default to the slurm launcher in chplenv
 - Revisit the options used
 - Add a numa option (that changes options appropriately) 
 - Allow walltime to be set from the launcher and not just env var 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22993 3a8e244f-b0f2-452b-bcba-4c88e055c3ca