Fix llvmAggregateGlobalOps instruction order problems
[reviewed by Tom]

Discovered situations in which this pass would create instructions using the
result of instructions that occur later. Fix that by changing the approach here
a little bit. We used to try to keep loads/stores in a sequence of loads/stores
that can be aggregated in the same order, so that intervening instructions have
their dependencies met. But the loads/stores all have compile-time constant
offsets from a base pointer - and the other instructions cannot have any effect
on memory. So it is safe to put all address computations before the sequence of
loads/stores and all instructions dependent on a loaded value after the
sequence.  This patch adds the described reordering and now outputs the
loads/stores in address order (rather than original order).  Also it was
necessary to not perform deletions of the original loads/stores until after the
function completed - since the first and last load/store needs to serve as
anchors for reordering.

Tested with 'CHPL_LLVM=none make check' and verified that with --llvm
--llvm-wide-opt --fast SSCA2 compiles and runs and
test/distributions/sungeun/dataPar1.chpl works as expected (SSCA2 and dataPar1
were showing this problem earlier).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22995 3a8e244f-b0f2-452b-bcba-4c88e055c3ca