Replace some occurrences of "inout" with "ref" or "out" module code.
Most standard module code was written before the "ref" intent was implemented,
and used "inout" instead.  We may assume that "ref" or "out" was intended --
particularly in external C code, where a deliberate copy-in/copy-out of a
value passed through a pointer would be rare.  

This patch deals with extern interfaces only.  In the Chapel compiler "ref",
"out" and "inout" are equivalent when comes to generating a call to an extern
function.  So the choice among the three is basically a matter of documenting
the interface.  

In the extern declarations related to QIO, I used the "out" intent where it
appeared to be the correct choice, and "ref" otherwise.  I left the non-extern
declarations alone.  Some of the implementations in qbuffer.c then did not meet
the requirement of guaranteed initialization for variables marked with "out"
intent, so I corrected them.

There was a comment in ChapelBase.chpl related to the implementation of
chpldev_refToString which seems obsolete, so I removed it.

Vass says: Technically, the change of inout -> out and the corresponding change
in qbuffer.c are not "bug fixes". However, they are minor and are committed
with Vass's approval.

[tested on release/examples/ (std)]
[tested on / (--no-local)]
[reviewed by vass]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22996 3a8e244f-b0f2-452b-bcba-4c88e055c3ca