Ensure fLinkStyle is set to the default after --static is passed on OS X
[Reviewed by Elliot]

fLinkStyle happens to be set to 1 (which is the same as LS_STATIC) inside
of proess_arg when --static is passed. So we need to set fLinkStyle back
to the default of LS_DEFAULT when we are on darwin to actually ignore the
flag.

This line should have been a part of commit r22998


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23001 3a8e244f-b0f2-452b-bcba-4c88e055c3ca